Add safety question when creating a large file

+++ generator.h
@@ -1,4 +1,4 @@
 #include <time.h>
 #include "std.h"
 
-int generate(int fd, EL_TYPE *buffer, size_t size,  int num_create);
+int generate(int fd, EL_TYPE *buffer, size_t size,  uint64_t num_create);
