+++ mergesort.h
@@ -2,4 +2,4 @@
 
 #include <sys/types.h>
 
-int start(int fd, EL_TYPE *buffer, size_t size, int num_threads);
+int start(int fd, EL_TYPE *buffer, size_t size, uint64_t num_threads);
